Step 1 of DotCloudification:
Add a "dotcloud.yml" file. For siwapp, we need a PHP service
and a MYSQL service. In the example dotcloud.yml here, we've
given them the names "www" and "mysql", respectively. We also
have to specify the type, and for the PHP service we designate
the directory "web" as our approot.